Added linkedlist template class + unit test

+++ include/bitops.hpp
@@ -1,8 +1,10 @@
 #pragma once
 #include <stdint.h>
 
-void set_bitpos(uint32_t& bitmask, const uint32_t& bitpos);
-void clr_bitpos(uint32_t& bitmask, const uint32_t& bitpos);
-uint32_t count_leading_zero(uint32_t& val);
-uint32_t find_bitpos_rightmost_zero(const uint32_t& val);
-bool check_is_pow_of_2(uint32_t& val);
+#include <cstddef>
+
+void set_bitpos(size_t& bitmask, const size_t& bitpos);
+void clr_bitpos(size_t& bitmask, const size_t& bitpos);
+size_t count_leading_zero(size_t& val);
+size_t find_bitpos_rightmost_zero(const size_t& val);
+bool check_is_pow_of_2(size_t& val);
